backend/pyproject.toml
[tool.poetry]
name = "fraud-detection-backend"
version = "0.1.0"
description = "AI-Powered Fraud Detection System"
authors = ["Admin <admin@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.27"
asyncpg = "^0.30.0"
pydantic = "^2.6.1"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "3.2.2" # Pin bcrypt to <4.0.0 to avoid passlib error
networkx = "^3.6"
langchain = "^0.3.0"
langgraph = "^0.2.20"
langchain-anthropic = "^0.2.0"
langchain-core = "^0.3.0"
langchain-postgres = "^0.0.14"
PyExifTool = "^0.5.6"
opencv-python-headless = "^4.9.0"
cryptography = "^42.0.0"
python-magic = "^0.4.27"
reportlab = "^4.0.0"
structlog = "^24.1.0"
prometheus-fastapi-instrumentator = "^7.0.0"
opentelemetry-api = "^1.24.0"
opentelemetry-sdk = "^1.24.0"
opentelemetry-instrumentation-fastapi = "^0.45b0"
opentelemetry-exporter-otlp = "^1.24.0"
 slowapi = "^0.1.9"
 redis = "^5.0.0"
 email-validator = "^2.3.0"
 qdrant-client = "^1.9.0"
 sentence-transformers = "^2.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.26.0"
pytest-asyncio = "^1.3.0"
aiosqlite = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
